bugfix(fix): renamed ApiError folder
